ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.12
pillow==11.0.0
rembg==2.0.60
torch==2.5.1
torchvision==0.20.1
transformers==4.48.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
